% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_cran_email.R
\name{fetch_cran_email}
\alias{fetch_cran_email}
\title{fetch the package submission email from cran in your gmail inbox}
\usage{
fetch_cran_email(pkg = ".")
}
\description{
fetch the package submission email from cran in your gmail inbox
}
