% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggqqplot.r
\name{ggqqplot}
\alias{ggqqplot}
\title{Make a q-q plot for model diagnostics}
\usage{
ggqqplot(model)
}
\arguments{
\item{model}{A model object, can be from lm() or glm() call I think.}
}
\description{
Make a q-q plot for model diagnostics
}
\examples{
\dontrun{
library(ggplot2)
data(mtcars)
mod <- lm(mpg ~ qsec, data=mtcars)
ggqqplot(mod)
}
}
