% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scott_release.R
\name{release}
\alias{release}
\alias{scott_release}
\alias{scott_release_submit}
\alias{scott_release_cleanup}
\alias{scott_release_git_tag}
\alias{scott_release_publish_release}
\title{Release to cran steps}
\usage{
scott_release()

scott_release_submit()

scott_release_cleanup()

scott_release_git_tag()

scott_release_publish_release()
}
\description{
Release to cran steps
}
\examples{
# scott_release()
}
