% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req.r
\name{req}
\alias{req}
\title{Easy require/library.}
\usage{
req(x)
}
\arguments{
\item{x}{Vector of quoted package names to load.}
}
\description{
Easy require/library.
}
\examples{
req(list("XML", "doMC", "plyr", "RCurl", "stringr"))
}
