% Generated by roxygen2 (4.0.0): do not edit by hand

\name{simbaltrees}
\alias{simbaltrees}
\title{Simulate a set of balanced and unbalanced trees.}
\usage{
  simbaltrees(tips = 10, metric, numtrees, cutlow, cuthigh)
}
\arguments{
  \item{tips}{}

  \item{metric}{Methods to use to generate trees, one of
  "colless", "beta", or gamma (see details). Defaults to
  "colless".}

  \item{numtrees}{Number of trees to produce. Defaults to
  10 trees.}

  \item{cutlow}{Value at which to filter trees on the low
  (e.g., unbalanced) side of the metric.}

  \item{cuthigh}{Value at which to filter trees on the high
  (e.g., balanced) side of the metric.}
}
\description{
  Could use this in asking questions aobut how phylogenetic
  tree balance influences ____ (madlib it).
}
\examples{
\dontrun{
temp <- simbaltrees(tips=10, metric="colless", numtrees=10, cutlow=-0.5, cuthigh=0.5)
unique(temp$model)
}
}

