% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scott_release.R
\name{release}
\alias{release}
\alias{scott_release}
\alias{scott_release_submit}
\alias{scott_release_cleanup}
\alias{scott_release_git_tag}
\alias{scott_release_publish_release}
\alias{scott_release_fetch_cran_email}
\title{Release to cran steps}
\usage{
scott_release(sleep = 1L)

scott_release_submit()

scott_release_cleanup()

scott_release_git_tag()

scott_release_publish_release()

scott_release_fetch_cran_email()
}
\arguments{
\item{sleep}{(integer) number of seconds to sleep before trying to
fetch the cran email with submission link. default: 1L}
}
\description{
Release to cran steps
}
\examples{
# scott_release()
}
